Edited user detail page to show videos uploaded by that user

+++ controllers/videoController.js
@@ -55,10 +55,10 @@ export const videoDetail = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
+    // console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
-    console.log(error);
+    // console.log(error);
     res.redirect(routes.home);
   }
 };
@@ -70,7 +70,13 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator != req.user.id) {
+      // console.log(video.creator);
+      // console.log(req.user.id);
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -95,7 +101,14 @@ export const deleteVideo = async (req, res) => {
     body: { fileUrl },
   } = req;
   try {
-    await Video.findOneAndDelete({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      // console.log(video.creator);
+      // console.log(req.user.id);
+      throw Error();
+    } else {
+      await Video.findOneAndDelete({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
